requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==24.3.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
defusedxml==0.7.1
deprecation==2.1.0
dj-database-url==2.3.0
dj-rest-auth==5.0.2
Django==5.0.3
django-allauth==0.61.1
django-cors-headers==4.3.1
django-filter==24.1
djangorestframework==3.14.0
frozenlist==1.5.0
gotrue==2.11.1
h11==0.14.0
h2==4.1.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
multidict==6.1.0
oauthlib==3.2.2
packaging==24.2
pillow==10.2.0
postgrest==0.19.1
propcache==0.2.1
psycopg2==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
realtime==2.1.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.2
storage3==0.11.1
StrEnum==0.4.15
supabase==2.11.0
supafunc==0.9.2
typing_extensions==4.12.2
urllib3==2.2.3
websockets==13.1
yarl==1.18.3
